Add files via upload
UsefulString: Takes username input from the keyboard and prints it as part of a receipt statement while considering taxes on the bill.
Usee: Works like the cintinue function in python. It primes the input and only starts to itirate after the character we want to omit is read.
Traffic: An elaborate conditional statement that considers whether a driver can proceed based on whuch traffic light is on and also considering whether a traffic cop is signaling to go or not.
TinyCalc: Uses switch statements to produce a functioning basic calculator with decent console output.
Tiny: a very basic adding machine. Very basic